requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==2.3.3
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pip==24.2
pool==0.1.2.dev0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dotenv==1.0.0
requests==2.31.0
rich==13.8.0
setuptools==74.1.1
SQLAlchemy==2.0.20
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
